[WIN-167][#59]🔨UserDefaultsService -> UserDomain
